+++ lib/cf-recvbuf.h
@@ -30,7 +30,7 @@
 
 CURLcode Curl_cf_recvbuf_add(struct Curl_easy *data,
                              struct connectdata *conn,
-                             int sockindex,
+                             int8_t sockindex,
                              const uint8_t *buf, size_t blen);
 
 extern struct Curl_cftype Curl_cft_recvbuf;
